+++ modules/develop/core.nix
@@ -7,6 +7,7 @@
     ./gemini-wrapper.nix
     ./gemini.nix
     ./opencode-wrapper.nix
+    ./paseo-wrapper.nix
     ./codex-wrapper.nix
   ];
 }