+++ modules/develop/paseo-wrapper.nix
@@ -1,0 +1,20 @@
+{ pkgs, inputs, ... }:
+
+let
+  agentTooling = import ./agent-tooling.nix {
+    inherit pkgs inputs;
+  };
+
+  paseo-cli = agentTooling.mkInstalledAgentWrapper {
+    name = "paseo";
+    binaryName = "paseo";
+    extraEnvironment = ''
+      export SSH_AUTH_SOCK="/run/user/1000/ssh-agent"
+    '';
+  };
+in
+{
+  environment.systemPackages = [
+    paseo-cli
+  ];
+}